pylibfdt: fix Python tests in Meson
Meson runs the run_tests.sh in the build directory, which means
pylibfdt_tests.py runs in the build directory, and pylibfdt_test.py
attempts to import libfdt.py at the relative path "../pylibfdt". But
the libfdt.py gets created in the source directory, which causes import
errors.

libfdt.py is in the source directory because Python distutils always
generates the SWIG output file in the same directory as the input file
libfdt.i. Distutils do not have an option to change the output to be in
a different directory[1]. So we work around the issue by copying
libfdt.i into the build directory.

[1] https://github.com/pypa/setuptools/blob/77c14a8e345a615c749fbac6345b07172cfc1e11/setuptools/_distutils/command/build_ext.py#L613-L617

Signed-off-by: Brandon Maier <brandon.maier@collins.com>