pylibfdt: move libfdt.py to build directory
During a Meson build, setup.py creates libfdt.py in the source
directory. When building under Meson we want to put build files into the
build directory to provide build isolation.

libfdt.py is in the source directory because distutils always tells swig
to create the output file in the same directory as the input file.
Distutils does not have an option to change the output to be in a
different directory[1]. So we work around the issue by copying libfdt.i
into the build directory.

[1] https://github.com/pypa/setuptools/blob/77c14a8e345a615c749fbac6345b07172cfc1e11/setuptools/_distutils/command/build_ext.py#L613-L617

Signed-off-by: Brandon Maier <brandon.maier@collins.com>